Ignore expressions that can not be pre-processed
This is because some expressions uses custom data models.
For example, an expression in a foreach token can use the temporary variable created in the loop, which can not be accessed by the root data resolver.
These cases generates exceptions but are not errors, so for now, just bypass exceptions and give a second chance to the token parser :-)